Move tid_to_pid() to tracelib.
